Reformat strings.
At first, I kept long strings all on one line - usually that's easier when
putting together sentences.  It turns out, though, that Ren'Py will strip out
the extra spaces, so hard-wrapping is a-ok.

And then when I figured that out, I was hard-wrapping to the same indent level,
which is just silly.

So fix both of those - hard wrap and indent in.